patch: fix finding of ctypes
The previous version worked with Snapcraft 7.3.2 because it was using
the Python interpreter from the base, which shouldn't be done for
classic snaps and fails in Snapcraft 7.4. With the inclusion of the
python3.10-minimal package, the bundled Python is found and so the
concatenation of ${SNAPCRAFT_PART_INSTALL} with the output of
`python3 -c 'import ctypes; print(ctypes.__file__)'` is no longer
needed - the result of that command is already inside the part's
install dir.